backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
PyPDF2
python-docx
requests
python-dotenv
transformers==4.40.0
torch>=2.0.0
